agent-server/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
google-genai
python-dotenv
pytest
pytest-asyncio
google-adk
